requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
whitenoise==6.9.0
python-decouple==3.8
requests==2.32.3